rename staking to stake
