+++ packages/wasm/crate/src/stake.rs
