+++ packages/wasm/src/stake.ts
